Fuzzer: Refactor into separate files
