+++ packages/tools/fuzzer/constants.ts
@@ -1,0 +1,5 @@
+import { dirname, join } from 'path';
+
+export const packagesDir = dirname(dirname(__dirname));
+export const cliDirectory = join(packagesDir, 'app-cli');
+
